Fixed problems with CLI authentication parameters.
Previously specifying a security database password in the CLI would
require a certificate nickname to be specified as well. While this
is correct for client certificate authentication, it caused a
problem for operations that do not authenticate against the server
such as client-init. The CLI has been modified to require a security
database password only if the nickname is specified for client
certificate authentication.

Similar changes have been made to require user password only if
the username is specified for basic authentication.

The CLI also has been modified to store all specified parameters
in the config object regardless of parameter validation.

The manual page has been modified accordingly.

Ticket #1125