Fix kra-connector-remove
The code to remove the connector from the pki CLI was found to be broken
because of invalid message type (partly due to void returns).

On uninstall, we need to remove the kra-connector from all relevant CA's
in the security domain.  The best way to do this is to keep kra-connector
info in LDAP, so that only one call is needed.  Until that change has
been made, we are adding a hack to remove the connector from all CA's
in the secutrity domain (if it exists).

Due to issues with proxy configurations, we will continue to use sslget
and a url-encoded-form version of the servlet.

In addition, it was found that when removing a KRA from a shared subsystem,
the updateDomainXML servlet was erroneously returning failure when it
was unsuccessful in removing a non-existent user from a group.

Ticket 1113