Added option to import client cert from CA.
A new option has been added to the client-cert-import command to
import a certificate from CA by specifying the serial number.

The client-cert-import has also been modified to get the nickname
of the certificate to import from the CLI argument. For backward
compatibility, if no argument is specified the CLI will try to
get the nickname from the authentication option (-n).

Ticket #1152