Fixed problem importing renewed system certificate.
Previously during clone installation if the PKCS12 file contains
both expired and renewed certificates the code might incorrectly
import the expired certificate instead of the renewed one, thus
failing the installation.

The code has been fixed to validate the certificates in the PKCS12
file such that only the valid ones will be imported into the clone.

https://fedorahosted.org/pki/ticket/1093