Fix-for-Bug-1170867-TPS-Installation-Failed
Fix now includes last review comments where we decided to consolidate 3 of the
ldif files: schema.ldif,database.ldif, and manager.ldif.

Each one of these 3 files contains the data needed for any subsystem for that file.
The subsystem specific files for these 3 go away in the source tree.

The first iteration of this fix was copying these 3 files into an undesirable directory.
This is no longer the case.

Extra code in the python installer allows one to establish a "file exclusion" callback to
keep a set of desired files from being copied when the installer does a directory copy.

All subsystems have been tested, including TPS with a brand new DS (which was the original reason for this fix),
and they appear to work fine.

Addressed further review comments:

1. Removed trailing whitespace instances from schema.ldif which had some.
2. Used pycharm to remove the few PEP violations I had previously added to the Python code.
3. Changed the format of the schema.ldif file to make all the entries use the same style.
Previously the TPS entries was using an all in one syntax. No more since now each entry is separate.
4. Changed the name of an argument in one of the new Python methods to get rid of a camelCase instance.

5. Tested everything to work as before, including basic TPS operations such as Format.

Fixed a method comment string and fixed some typos.