Refactored CRMFPopClient.
The CRMFPopClient has been refactored such that it is easier
to understand and reuse. The code has been fixed such that it
can read a normal PEM transport certificate. It also has been
fixed to parse the request submission result properly.

The client-cert-request CLI command was modified to support CRMF
requests.

The MainCLI and ClientConfig were modified to accept a security
token name.

The pki_java_command_wrapper.in was modified to include the Apache
Commons IO library.

https://fedorahosted.org/pki/ticket/1074