Fixed pylint report.
Previously pylint report was saved it into a file which may not be
accessible on a build system. The pylint-build-scan.sh has been
changed to display the report so it will appear in the build log.

The pylint configuration has also been modified to disable C and R
messages by default. This way when other errors or warnings occur
the build will fail without having to check for specific codes.

Some Python codes have been modified to reduce the number of pylint
warnings.

https://fedorahosted.org/pki/ticket/703