Added support for Tomcat 8.
The Dogtag code has been modified to support both Tomcat 7 and 8.
All files depending on a specific Tomcat version are now stored
in separate folders. The build scripts have been modified to use
the proper folder for the target platform. The tomcatjss
dependency has been updated as well.

The upgrade script will be added in a separate patch.

https://fedorahosted.org/pki/ticket/1264