Store groups on AuthToken and update group evaluator
Update the UidPwdDirAuthentication plugin to retrieve all the user's
groups from a directory and store them on the AuthToken.

Also update the group evaluator to match against all the groups
stored in the AuthToken.  The "gid" and "groups" are merged into a
single collection, if the ACL operation is "=" the collection is
checked under disjunction, and if the operation is "!=", then
conjunction.

Fixes https://fedorahosted.org/pki/ticket/1174