Patches to get nuxwdog working with systemd
This patch adds some new unit files and targets for starting instances
with nuxwdog, as well as logic within the pki-server nuxwdog module to
switch to/from the old and new systemd unit files.

It also corrects some issues found in additional testing of the nuxwdog
change scripts.

To use nuxwdog to start the instance, a user needs to do the following:

1. Create an instance normally.
2. Run: pki-server instance-nuxwdog-enable <instance_name>
3. Start the instance using:
   systemctl start pki-tomcatd-nuxwdog@<instance_name>.service

To revert the instance, simply do the following:

1. Run: pki-server instance-nuxwdog-disable <instance_name>
2. Start the instance using:
   systemctl start pki-tomcatd@<instance_name>.service