Fix #1351 pki securitydomain-get-install-token fails when run with caadmin user.
The short term solution to this problem was to remove the man page information and all references to the command line module reponsible for this issue.

The installer already has an alternative method to remove a subsystem from the security domain list. We now assume the alternate method and don't even try to find the token at this point.

A user at the command line of the pki command will no longer be able to attempt this as well.

Tested this to verify that the man page for the "securtydomain" command no longer mentions or documents the "get-install-token" variant. Tested to verify that this command can't be manually called from the command line using "pki". This attempt results in an "unknown module". Tested by installing and uninstalling a subsytem. The security domain was kept up to date as expected for each install over remove attempted.