Fixed key archival problem in CLI with separate KRA instance.
The CLI has been modified such that when enrolling a certificate
with key archival it will obtain the transport certificate from
the CA instead of KRA because the KRA may not reside on the same
instance. The CA REST service has been modified such that it will
obtain the transport certificate from the KRA connector.

https://fedorahosted.org/pki/ticket/1384