Cleaned up links in main page.
The ROOT's index.jsp has been modified to show the links to all
subsystems installed on the instance. When opened, it will show
the services provided by the subsystem.

The pkispawn output has been modified to show the subsystem URL
more consistently:
   https://<hostname>:<port>/<subsystem>

In all subsystems except TPS the page will redirect to:
   https://<hostname>:<port>/<subsystem>/services