Fixed user-cert-add --serial with remote CA.
The user-cert-add command has been modified to ask the user for
the CA server URI if the CA is not available locally.

A new SubsystemClient.exists() method has been added to check
whether a subsystem is deployed on the target instance.

The SubsystemCLI has been modified to call logout() only if
the operation is executed successfully.

The certificate approval callback class has been refactored out
of PKIConnection into a separate class to clean up circular
dependency with PKIClient.

https://fedorahosted.org/pki/ticket/1448