Simplify exception handling in pkihelper
Several methods except OSError before they except shutil.Error. In
Python 3 the second except clause will be ignored because in Python 3
shutil.Error is a subclass of OSError.

The body of the except clauses only differs in the logging message. A
single except clause with an isinstance() check has the same effect.