Handle JSON decode error in handle_exceptions()
pki.handle_exceptions() raises a JSON decode exception when the body of
the HTTPException is not a valid JSON string. The JSON exception hides
the true error message.

The patch also fixes a bug in PKIException.from_json(). The code and
ClassName attribute are now correctly set. Finally we have our first
unit test.

https://fedorahosted.org/pki/ticket/1488
https://fedorahosted.org/freeipa/ticket/5129