Fix exception when talking to dogtag 9 systems
When getting a token from the security domain for a Dogtag 9
system, we first attempt to reach the REST interfaces.  When this
fails (with 404 exception), we catch the exception and try the
old interfaces.

The exception being thrown has been changed from the deprecated
ClientResponseFailure to being wrapped in a PKIException, so the
code catching the exception needs to be modified accordingly.

Ticket 1495