Fix code to add replicationdb password unless already present
The replicationdb password is an instance parameter and should
be created by the first subsystem in the instance.  This should
happen independantly of whether replication is being set up
in case it is needed to set up replication (as a master) later.

Related to Ticket 1414

(cherry picked from commit 8a002e091841ceb52346f121d75db80e78735af2)