Python packaging of PKI client library
A new setup.py in base/common/python makes it possible to bundle
the pki client library and upload it on PyPI. The setup.py in the root
directory is only used for tox and testing. It's a cleaner and less
fragile approach than to support two different build flavors with one
setup.py

The 'release' alias from setup.cfg creates and uploads a source
distribution and an universal wheel:

  $ sudo yum install python-wheel python-setuptools
  $ cd base/common/python
  $ python setup.py release

The 'packages' alias just creates the source distribution and wheel:

  $ python setup.py packages

The version number is taken from the Version and Release fields of
pki-core.spec.