Added automatic Tomcat migration.
The pki-core.spec has been modified to execute pki-server migrate
when the package is installed. This way when upgrading from F22 to
F23 all PKI instances will be migrated automatically to Tomcat 8.

The pki-server migrate command has been modified such that if there
is no specific Tomcat version specified it will use the current
Tomcat version.

The top attribute in the CLI class was not functioning properly,
so it has been replaced with get_top_module() method.

The getopt() invocations in pki-server subcommands have been
replaced with gnu_getopt() to allow intermixing options and
arguments.

https://fedorahosted.org/pki/ticket/1310