Block startup until initial profile load completed
It is possible for the CMS getStatus resource to indicate that CMS
is ready when the initial loading of profiles (which is performed by
another thread) is not complete.  During startup, wait for the
initial loading of profiles to complete before continuing.

Fixes: https://fedorahosted.org/pki/ticket/1702
(cherry picked from commit e8a1d9cbbefb2988092e96b13d0b13254c92d1b2)