Ticket #1963 CRL generation enters loop when CA loses connection to netHSM. This patch makes a low risk attempt to slow down the loop that could be caused by an unexpected exception caused by the unavailability of a dependant component (e.g. HSM, LDAP) in the middle of CRL generation/update.
New configuration parameters are:
ca.crl.MasterCRL.unexpectedExceptionWaitTime
  - the wait time in minutes; default is 30
  - normally you want it to be less than ca.crl.MasterCRL.autoUpdateInterval
    and ca.crl.MasterCRL.cacheUpdateInterval
ca.crl.MasterCRL.unexpectedExceptionLoopMax
  - the max number of tries allowed before the slow down mechanism kicks in;
    default is 10

When such unexpected failure happens, a loop counter is kept and checked
against the unexpectedExceptionLoopMax.  If the loop counter exceeds the
unexpectedExceptionLoopMax, then the current time is checked against the
time of the failure, where the time lapse must exceed the
unexpectedExceptionWaitTime to trigger a delay.  This delay is the
counter measure to mitigate the amount of log messages that could flood
the log(s).
The delay is calcuated like this:
    waitTime = mUnexpectedExceptionWaitTime - (now - timeOfUnexpectedFailure);