Lightweight CAs: set DN based on data from LDAP
When initialising a lightweight CA, if we do not have the signing
cert and key in the NSSDB yet, we do not initialise the DN.  This
causes NPE in other code that expects getX500Name() to return a
value, e.g. REST API to list or show CA.

To work around this, when loading lightweight CAs set the DN based
on the 'authorityDN' value stored in its LDAP entry.

Part of: https://fedorahosted.org/pki/ticket/1625