Updated pki pkcs12-export CLI.
For consistency the pki pkcs12-export has been modified to
overwrite the PKCS #12 output file by default. A new option has
been added to append the exported certificates and keys into the
output file if the file already exists.

The same option has been added to the The pki-server
instance-cert-export and subsystem-cert-export commands.

https://fedorahosted.org/pki/ticket/1736