Realm: allow auth instances to support multiple realms
In practice, most folks will use something like DirAclAuthz
to manage their realm.  Rather than requiring a new authz plugin
for each realm, we allow the authz plugin to support multiple
realms (as a comma separated list).

For the Acl plugins in particular, we expand the authorize call
to allow the caller to pass in the realm as well as the resource
and operation.  The resource queried would then be constructed on
the fly as realm.resource

Examples will be provided in the wiki page.

Trac Ticket 2041