Fix problem in creating certificate requests
Some incorrect code was added to request processing
in the realm patches.  In the request LDAP modification code,
if the realm was not present, we added a modification to remove the
realm attribute.

Unfortunately, if the realm was not present to begin with, this resulted
in LDAP returning a "No Such Attribute (16)" error, causing all kinds
of requests - including certificate requests to fail to be submitted.

At this point, we do not permit users to change the realm of a request.
Therefore, there is no reason to remove the realm.  If we ever need
to do this in future, we'll have to be smarter about it.