Renamed token status TEMP_LOST to SUSPENDED.
The token status TEMP_LOST has been renamed to SUSPENDED such that
it can be used more general contexts.

To simplify the transition, the CLIs and the REST API will continue
to accept TEMP_LOST but it will be converted internally into
SUSPENDED and a deprecation warning will be generated.

https://fedorahosted.org/pki/ticket/2286