Enhance tkstool for capabilities and security
The key is now generated with the flags needed to keep the data from being displayed
with simple tools such as symkeyutil.

As per cfu's instructions,
I was able to test this with the nethsm only.

I also was able to make the key des3 and everything works fine with the master key.
This will help all the warnings we get about insecure des2 keys.

If there is a problem with luna, we can file another ticket.
Also there could be a built in tool for luna to generate keys such as is present on hsm.